==> notebooks/05_fusion_enrichissement.ipynb <==
# Etape 2.2 : Fusion et enrichissement

## Charger les consommations nettoyees (depuis Parquet)

import pandas as pd
import warnings
warnings.filterwarnings('ignore')

# Configuration affichage
pd.set_option('display.max_columns', None)
pd.set_option('display.width', None)

# Chemins
DATA_DIR = "../data"
OUTPUT_DIR = "../data/output"
PARQUET_DIR = "../data/output/consommations_clean"


# chargement des données
df_consommations = pd.read_parquet(PARQUET_DIR, engine='fastparquet')
print(f"Données chargées : {len(df_consommations):,} lignes")


df_consommations.head(5)

### - Fusionner avec les donnees meteo (sur commune et timestamp arrondi a l'heure)

df_meteo = pd.read_csv("../data/output/meteo_clean.csv" , sep=",")
df_meteo.head(5)
df_join_meteo = pd.merge(
    df_consommations, 
    df_meteo, 
    on=["commune", "timestamp", "month"], 
    how="inner"
)

df_join_meteo.head(5)

### - Fusionner avec le referentiel batiments

df_batiments = pd.read_csv("../data/batiments.csv" , sep=",")
df_batiments.head(5)
df_join_batiments = df_join_meteo.merge(
    df_batiments,
    on=['commune','batiment_id'],
    how='left'
)

df_join_batiments.head(5)

### - Fusionner avec les tarifs pour calculer le cout financier

df_tarifs = pd.read_csv("../data/tarifs_energie.csv" , sep=",")
df_tarifs.head(5)

df_join_batiments['date'] = pd.to_datetime(df_join_batiments['date'])
df_tarifs['date_debut'] = pd.to_datetime(df_tarifs['date_debut'])
df_tarifs['date_fin'] = pd.to_datetime(df_tarifs['date_fin'])

df_final = df_join_batiments.merge(
    df_tarifs,
    on='type_energie',
    how='left'
)

# df_final = df_join_batiments.merge(
#     df_tarifs,
#     on=['commune','batiment_id'],
#     how='left'
# )

df_final.head(5)


### Creer des features derivees :


  - Consommation par occupant


df_final['consommation_par_occupant'] = df_final['consommation_clean'] / df_final['nb_occupants_moyen']

  - Consommation par m2

df_final['consommation_par_m2'] = df_final['consommation_clean'] / df_final['surface_m2']

  - Cout journalier, mensuel, annuel

df_final['cout_instantane'] = df_final['consommation_clean'] * df_final['tarif_unitaire']
costs = df_final.groupby('batiment_id')['cout_instantane'].agg(['sum', 'mean'])

  - Indice de performance energetique (IPE)

conso_annuelle = df_final.groupby('batiment_id')['consommation_clean'].transform('sum')
df_final['IPE'] = conso_annuelle / df_final['surface_m2']

  - Ecart a la moyenne de la categorie

df_final['moyenne_type'] = df_final.groupby('type')['consommation_par_m2'].transform('mean')

df_final['ecart_moyenne_categorie_pct'] = ((df_final['consommation_par_m2'] - df_final['moyenne_type']) / df_final['moyenne_type']) * 100

df_final.head(5)

# Exporter le fichier
df_final.to_csv('../data/output/consommations_enrichies.csv', index=False)
